renamed lib/form to lib/indexDB
