+++ src/lib/index.ts
@@ -2,4 +2,4 @@ export * from "./dates";
 export * from "./icons";
 export * from "./tanstack";
 export * from "./utils";
-export * from "./form";
+export * from "./indexDB";
